requirements.txt
# requirements.txt
MetaTrader5
pandas
numpy
pytz
python-dotenv
